Path for Build-Exe - added path for build-exe - added swi-prolog
Former-commit-id: cf2cf630aac0ce4234bc523f6575a06e0b4609e7

+++ TrafficLightControl/Assets/Scripts/Logger/UnityLogger.cs
@@ -14,7 +14,12 @@ public class UnityLogger : MonoBehaviour {
     // Use this for initialization
     void Start () {
         _prologQueue = new Queue<string>();
-        _swProlog = new StreamWriter(@".\Assets\Prolog\prolog.log", true);
+
+        if (Directory.Exists(@".\Assets\Prolog"))
+            _swProlog = new StreamWriter(@".\Assets\Prolog\prolog.log", true);
+        else
+            _swProlog = new StreamWriter(@".\Prolog\prolog.log", true);
+
     }
 
     void OnApplicationQuit()
